requirements.txt
altgraph==0.17.4
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.56.0
fsspec==2025.2.0
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
macholib==1.16.3
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
seaborn==0.13.2
setuptools==77.0.3
six==1.17.0
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.78
ultralytics-thop==2.0.14
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
